chore: complete repository backup

+++ server/app/account/models.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, Float, String
+from sqlalchemy import Boolean, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -13,8 +13,8 @@ class Account(Base):
     id: Mapped[str] = mapped_column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     type: Mapped[str] = mapped_column(String(20), nullable=False)
-    balance: Mapped[float] = mapped_column(Float, default=0.0)
-    initial_balance: Mapped[float] = mapped_column(Float, default=0.0)
+    balance: Mapped[float] = mapped_column(Numeric(12, 2), default=0.0)
+    initial_balance: Mapped[float] = mapped_column(Numeric(12, 2), default=0.0)
     icon: Mapped[str] = mapped_column(String(50), default="")
     color: Mapped[str] = mapped_column(String(20), default="")
     description: Mapped[str] = mapped_column(String(500), default="")
