+++ server/app/recurring_expense/models.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime, timezone
 from typing import Optional
 
-from sqlalchemy import Float, Integer, String
+from sqlalchemy import Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -13,7 +13,7 @@ class RecurringExpense(Base):
 
     id: Mapped[str] = mapped_column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     name: Mapped[str] = mapped_column(String(200), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
     day_of_month: Mapped[int] = mapped_column(Integer, nullable=False)  # 1-31
     category_id: Mapped[Optional[str]] = mapped_column(String(36), nullable=True)
     account_id: Mapped[Optional[str]] = mapped_column(String(36), nullable=True)
